Migrate fjs/effects/module.ts and fjs/types/result/module.ts to .mjs
Plain .ts implementation modules (not .f.ts) — asyncRun, and
tryCatch/asyncTryCatch — following the module.ts -> module.mjs stage-1
mapping. Updates fjs/effects/node/module.ts,
fjs/effects/node/memory/module.ts, and
fjs/effects/node/memory/proof.ts's imports accordingly.
