--- fjs/effects/module.ts
@@ -1,16 +1,0 @@
-import { match } from "./module.f.mjs"
-import type { Effect, Operation, ToAsyncOperationMap } from "./types.ts"
-
-export const asyncRun =
-    <O extends Operation>(map: ToAsyncOperationMap<O>) =>
-    async<T, E extends Effect<O, T>>(effect: Effect<O, T>): Promise<T> =>
-{
-    const next = match(map)
-    while (true) {
-        const r = next(effect)
-        if (r[0] === 'done') {
-            return r[1]
-        }
-        effect = r[2](await r[1])
-    }
-}
