--- fjs/types/result/module.ts
@@ -1,24 +1,0 @@
-import type { Result } from './types.ts'
-import { ok, error } from './module.f.mjs'
-
-export const tryCatch
-    : <T>(f: () => T) => Result<T, unknown>
-    = f => {
-        // Side effect: `try catch` is not allowed in FunctionalScript.
-        try {
-            return ok(f())
-        } catch (e) {
-            return error(e)
-        }
-    }
-
-export const asyncTryCatch
-    : <T>(f: () => Promise<T>) => Promise<Result<T, unknown>>
-    = async f => {
-        // Side effect: `try catch` is not allowed in FunctionalScript.
-        try {
-            return ok(await f())
-        } catch (e) {
-            return error(e)
-        }
-    }
